+++ doc/dev/api/v1.md
@@ -34,8 +34,8 @@ The `OctocatalogDiff::API::V1::Diff` object represents a difference between two 
 
 [Read more about the `OctocatalogDiff::API::V1::Diff` object](/doc/dev/api/v1/objects/diff.md)
 
-#### OctocatalogDiff::API::V1::FactOverride
+#### OctocatalogDiff::API::V1::Override
 
-The `OctocatalogDiff::API::V1::FactOverride` object represents a user-supplied fact that will be used when compiling a catalog.
+The `OctocatalogDiff::API::V1::Override` object represents a user-supplied fact or ENC parameter that will be used when compiling a catalog.
 
-[Read more about the `OctocatalogDiff::API::V1::FactOverride` object](/doc/dev/api/v1/objects/fact_override.md)
+[Read more about the `OctocatalogDiff::API::V1::Override` object](/doc/dev/api/v1/objects/override.md)
